% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoimage.R
\name{autoimage}
\alias{autoimage}
\title{Automatic facetting of multiple projected images}
\usage{
autoimage(x, y, z, legend = "horizontal", proj = "none", proj.args,
  lratio = 0.2, common.legend = TRUE, map = "none", size, outer.title,
  ...)
}
\arguments{
\item{x}{Locations of grid points at which the values in \code{z} 
are measured.  The values must be finite and non-missing.  These 
arguments can be either vectors, in which case the values must be in (strictly) ascending order.  If an irregular grid is to be used, then \code{x} and \code{y} should be numeric matrices having the same number of rows and columns as \code{z}.  If these arguments are not provided, equally spaced values from 0 to 1 are used by default. If \code{x} is a list, its components \code{x$x} and \code{x$y} are used for \code{x} and \code{y}, respectively. If the list has component \code{x$z}, this is used for \code{z}.}

\item{y}{Locations of grid points at which the values in \code{z} 
are measured.  The values must be finite and non-missing.  These 
arguments can be either vectors, in which case the values must be in (strictly) ascending order.  If an irregular grid is to be used, then \code{x} and \code{y} should be numeric matrices having the same number of rows and columns as \code{z}.  If these arguments are not provided, equally spaced values from 0 to 1 are used by default. If \code{x} is a list, its components \code{x$x} and \code{x$y} are used for \code{x} and \code{y}, respectively. If the list has component \code{x$z}, this is used for \code{z}.}

\item{z}{A numeric or logical matrix containing the values to be plotted (NAs are allowed). If multiple images are to be plotted, a numeric array can be provided instead.  The third dimension of the array indicates the number of images that should be plotted.  Note that \code{x} can be used instead of \code{z} for convenience.}

\item{legend}{A character string indicating where the color scale
should be placed.  The default is \code{"horizontal"}.  The other
valid options are \code{"none"} and \code{"vertical"}.}

\item{proj}{A character string indicating what projection should be
used for the included \code{x} and \code{y} coordinates.  The
default is \code{"none"}.  The other valid choices correspond to the
\code{"projection"} argument in the \code{\link[mapproj]{mapproject}} 
function, which is used for the projection.}

\item{proj.args}{A named list with arguments \code{parameters} and
\code{orientation} corresponding to the arguments of the same name
in the \code{\link[mapproj]{mapproject}} function.}

\item{lratio}{A numeric value indicating the ratio of the 
smaller dimension of the legend scale to the width of 
the image.  Default is \code{lratio = 0.2}.}

\item{common.legend}{A logical value indicating whether a common legend scale should be used for all images provided in the \code{z} array.  Default is \code{TRUE}.  If \code{FALSE}, a separate legend is used for each image.}

\item{map}{The name of map to draw on the image.
Default is \code{"none"}.  Other options include
\code{"world"}, \code{"usa"}, \code{"state"}, \code{"county"}, 
\code{"france"}, \code{"nz"} (New Zealand), and \code{"world2"}, 
all from the \code{maps} package.}

\item{size}{A vector of length two indicating the number of rows and columns that should be used for the series of image data in \code{z}.  Note that \code{prod(size)} must match the length of the third dimension of \code{z} (if it is an array), or \code{c(1, 1)} if \code{z} is a matrix.}

\item{outer.title}{A title related to all of the images that is plotted in the outer margin of the figure.}

\item{...}{Additional arguments passed to the \code{\link[graphics]{image}} or 
\code{\link[fields]{poly.image}} functions.  e.g., \code{xlab}, \code{ylab}, 
\code{xlim}, \code{ylim}, \code{zlim}, etc.}
}
\description{
\code{autoimage} plots a sequence of images (with possibly 
projected coordinates) while also automatically plotting a 
color scale matching the image colors to the values of \code{z}.  
Many options are available for legend customization.  The coordinates
can be irregularly spaced, on a regular grid, or on an irregular 
grid.  \code{z} can be a numeric vector, matrix, or array, depending
on the context.
}
\details{
The \code{\link[mapproj]{mapproject}} function is used to project 
the \code{x} and \code{y} coordinates when \code{proj != "none"}.

If multiple images are to be plotted (i.e., if \code{z} is an array), 
then the \code{main} argument can be a vector with length matching 
\code{dim(z)[3]}, and each successive element of the vector will 
be used to add a title to each successive image plotted.  
See the Examples.

Additionally, if \code{common.legend = FALSE}, then separate limits 
for the z-axis of each image can be provided as a list.  
Specifically, if \code{dim(z)[3] == k}, then \code{zlim} should 
be a list of length \code{k}, and each element of the list should 
be a 2-dimensional vector providing the lower and upper limit, 
respectively, of the legend for each image.  Alternatively, if 
\code{zlim} is a list of length \code{k}, then \code{common.legend}
is set to \code{FALSE}.

The range of \code{zlim} is cut into \eqn{n} partitions, 
where \code{n} is the length of \code{col}.

Note that the more images that are plotted simulataneously, 
the smaller one typically wants \code{lratio} to be.

The multiple plots are constructed using the 
\code{\link[autoimage]{autolayout}} function, which 
is incompatible with the \code{mfrow} and \code{mfcol} arguments 
in the \code{\link[graphics]{par}} function and is also 
incompatible with the \code{\link[graphics]{split.screen}} function.
}
\examples{
data(narccap)
# restructure data for 2 images
tasmax2 = tasmax[,,1:2]

# plot irregularly gridded images with separate legends
# and usa border
autoimage(lon, lat, tasmax2, common.legend = FALSE, map = "usa")

# plot irregularly gridded images with common legend and world lines
# customize world lines
# add and customize title
autoimage(lon, lat, tasmax2, map = "world", 
          lines.args = list(col = "white", lwd = 2),
          outer.title = "Maximum Daily Surface Air Temperature (K)",
          mtext.args = list(col = "blue", cex = 2))

# plot irregularly-spaced responsed as images with separate legends
# and county borders.  Add observed data locations with custom point
# options
data(co, package = "gear")
autoimage(co$lon, co$lat, co[,c("Al", "Ca")], common.legend = FALSE, 
          map = "county", main = c("Aluminum", "Cadmium"),
          points = list(x = co$easting, y = co$northing),
          points.args = list(pch = 20, col = "white"))

# customize margins and lratio for large plot
# also use projection
# specify manual lines (though in this case it is the same as using map = "world")
data(worldMapEnv, package = "maps")
worldpoly <- maps::map("world", plot = FALSE)
par(mar = c(1.1, 4.1, 2.1, 1.1))
autoimage(lon, lat, tasmax, lines = worldpoly, 
          proj = "bonne", proj.args = list(parameters = 40),
          main = c("day 1", "day 2", "day 3", "day 4", "day 5"),
          ylab = "",
          axes = FALSE,
          lratio = 0.5)
}
\seealso{
\code{\link[autoimage]{pimage}}
}

