% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoimage2.R
\name{autoimage2}
\alias{autoimage2}
\title{Automatic facetting of multiple projected images}
\usage{
autoimage2(x, y, z, legend = "horizontal", proj = "none", proj.args,
  lratio = 0.2, common.legend = TRUE, size, outer.title, ...)
}
\arguments{
\item{x}{Locations of grid points at which the values in \code{z} 
are measured.  The values must be finite and non-missing.  These 
arguments can be either vectors, in which case the values must be in (strictly) ascending order.  If an irregular grid is to be used, then \code{x} and \code{y} should be numeric matrices having the same number of rows and columns as \code{z}.  If these arguments are not provided, equally spaced values from 0 to 1 are used by default. If \code{x} is a list, its components \code{x$x} and \code{x$y} are used for \code{x} and \code{y}, respectively. If the list has component \code{x$z}, this is used for \code{z}.}

\item{y}{Locations of grid points at which the values in \code{z} 
are measured.  The values must be finite and non-missing.  These 
arguments can be either vectors, in which case the values must be in (strictly) ascending order.  If an irregular grid is to be used, then \code{x} and \code{y} should be numeric matrices having the same number of rows and columns as \code{z}.  If these arguments are not provided, equally spaced values from 0 to 1 are used by default. If \code{x} is a list, its components \code{x$x} and \code{x$y} are used for \code{x} and \code{y}, respectively. If the list has component \code{x$z}, this is used for \code{z}.}

\item{z}{A numeric or logical matrix containing the values to be plotted (NAs are allowed). If multiple images are to be plotted, a numeric array can be provided instead.  The third dimension of the array indicates the number of images that should be plotted.  Note that \code{x} can be used instead of \code{z} for convenience.}

\item{legend}{A character string indicating where the color scale
should be placed.  The default is \code{"horizontal"}.  The other
valid options are \code{"none"} and \code{"vertical"}.}

\item{proj}{A character string indicating what projection should be
used for the included \code{x} and \code{y} coordinates.  The
default is \code{"none"}.  The other valid choices correspond to the
\code{"projection"} argument in the \code{\link[mapproj]{mapproject}} 
function, which is used for the projection.}

\item{proj.args}{A named list with arguments \code{parameters} and
\code{orientation} corresponding to the arguments of the same name
in the \code{\link[mapproj]{mapproject}} function.}

\item{lratio}{A numeric value indicating the ratio of the 
smaller dimension of the legend scale to the width of 
the image.  Default is \code{lratio = 0.2}.}

\item{common.legend}{A logical value indicating whether a common legend scale should be used for all images provided in the \code{z} array.  Default is \code{TRUE}.  If \code{FALSE}, a separate legend is used for each image.}

\item{size}{A vector of length two indicating the number of rows and columns that should be used for the series of image data in \code{z}.  Note that \code{prod(size)} must match the length of the third dimension of \code{z} (if it is an array), or \code{c(1, 1)} if \code{z} is a matrix.}

\item{outer.title}{A title related to all of the images that is plotted in the outer margin of the figure.}

\item{...}{Additional arguments passed to the \code{\link[graphics]{image}} or \code{\link[fields]{poly.image}} functions.  e.g., \code{xlab}, \code{ylab}, \code{xlim}, \code{ylim}, \code{zlim}, etc.}
}
\description{
\code{autoimage} plots a sequence of images (with possibly 
projected coordinates) while also automatically plotting a 
color scale matching the image colors to the values of \code{z}.  
Many options are available for legend customization.  The coordinates
can be irregularly spaced, on a regular grid, or on an irregular 
grid.  \code{z} can be a numeric vector, matrix, or array, depending
on the context.
}
\details{
The \code{\link[mapproj]{mapproject}} function is used to project 
the \code{x} and \code{y} coordinates when \code{proj != "none"}.

If multiple images are to be plotted (i.e., if \code{z} is an array), 
then the \code{main} argument can be a vector with length matching 
\code{dim(z)[3]}, and each successive element of the vector will 
be used to add a title to each successive image plotted.  
See the Examples.

Additionally, if \code{common.legend = FALSE}, then separate limits 
for the z-axis of each image can be provided as a list.  
Specifically, if \code{dim(z)[3] == k}, then \code{zlim} should 
be a list of length \code{k}, and each element of the list should 
be a 2-dimensional vector providing the lower and upper limit, 
respectively, of the legend for each image.  Alternatively, if 
\code{zlim} is a list of length \code{k}, then \code{common.legend}
is set to \code{FALSE}.

The range of \code{zlim} is cut into \eqn{n} partitions, 
where \code{n} is the length of \code{col}.

Note that the more images that are plotted simulataneously, 
the smaller one typically wants \code{lratio} to be.

The multiple plots are constructed using the 
\code{\link[autoimage]{autolayout}} function, which 
is incompatible with the \code{mfrow} and \code{mfcol} arguments 
in the \code{\link[graphics]{par}} function and is also 
incompatible with the \code{\link[graphics]{split.screen}} function.
}
\examples{
# Example from image function documentation
x <- y <- seq(-4*pi, 4*pi, len = 27)
r <- sqrt(outer(x^2, y^2, "+"))
z <- cos(r^2)*exp(-r/6)
image(z, col  = gray((0:32)/32))
autoimage(z, col  = gray((0:32)/32), legend = FALSE)

# now with legend
autoimage(z, col  = gray((0:32)/32))
autoimage(z, col  = gray((0:32)/32), horizontal = FALSE)

# add some customization
autoimage(x, y, z, xlab = "x1", ylab = "y1", main = "Math is beautiful ...")

# now do some examples with multiple images
z2 <- cos(r^2/2)*exp(-r/3)
z3 <- cos(r^2/2)*exp(-r/6)
z4 <- cos(r^2/3)*exp(-r/5)
zarray <- abind::abind(z, z2, z3, z4, along = 3)

# multiple images with common scale, separate titles
autoimage(x, y, zarray, main = letters[1:4], size = c(2, 2),
          mratio = 3, mmar = c(4.1, 4.1, 2.1, 2.1))
# change the orientation of the scale
autoimage(x, y, zarray, main = letters[1:4], size = c(2, 2),
          mratio = 4, mmar = c(4.1, 4.1, 2.1, 2.1),
          horizontal = FALSE)

# add overall title to plots
autoimage(x, y, zarray, main = letters[1:4], size = c(2, 2),
          mratio = 4, mmar = c(4.1, 4.1, 2.1, 2.1),
          horizontal = FALSE, 
          outer.title = "Interesting images with colored title",
          mtext.args = list(col = "blue"))
# multiple images with separate legends
autoimage(x, y, zarray, size = c(2, 2),
          mratio = 4, mmar = c(4.1, 4.1, 2.1, 2.1),
          horizontal = FALSE, common.legend = FALSE,
          outer.title = "Interesting images")

# do some examples with an irregular grid
# load data from fields package
data(narccap)
# restructure data for 2 images
tasmax2 = tasmax[,,1:2]
tasmax4 = tasmax[,,1:4]

# plot irregularly gridded images
autoimage(lon, lat, tasmax2, col = fields::tim.colors(12), size = c(1, 2))

# Do the same plot, but with a projection.
# Notice that the axis scales seem off because of the projection
autoimage(lon, lat, tasmax2, col = fields::tim.colors(12), size = c(1, 2),
          project = TRUE,
          project.args = list(projection = "albers", parameters = c(33, 45)))
# compare the axes for the projected coordinates to the correct references lines using map.grid.
autoimage(lon, lat, tasmax2, col = fields::tim.colors(12), size = c(1, 2),
          project = TRUE, map.grid = TRUE,
          project.args = list(projection = "albers", parameters = c(33, 45)),
          grid.args = list(col = "black", nx = 5, ny = 5))
# turn axes off
autoimage(lon, lat, tasmax2, col = fields::tim.colors(12), size = c(1, 2),
          project = TRUE, map.grid = TRUE, axes = FALSE, xlab = "", ylab = "",
          project.args = list(projection = "albers", parameters = c(33, 45)),
          grid.args = list(col = "black", nx = 5, ny = 5))

# more images in a plot.  Need to change mratio
autoimage(lon, lat, tasmax4, col = fields::tim.colors(12), size = c(2, 2), 
          horizontal = FALSE, mratio = 4)
autoimage(lon, lat, tasmax4, col = fields::tim.colors(12), size = c(2, 2), mratio = 4)

# add a nice polygon to the images
library(maps) # need to get world map
# get the polygon for the world from the maps package
worldpoly = map("world", plot = FALSE)
autoimage(lon, lat, tasmax2, size = c(1, 2), project = TRUE, 
          project.args = list(projection = "albers", parameters = c(33, 45)),
          map.poly = worldpoly, axes = FALSE,
          mmar = c(0.5, 0.5, 0.5, 0.5), legend.mar = c(2, 0.5, 0.5, 0.5), 
          map.points = list(x = -108.529, y = 43.33091), 
          points.args = list(pch = 4, cex = 2, lwd = 2),
          outer.title = "NARCCAP output")
}
\seealso{
\code{\link[autoimage]{pimage}}
}

