% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.plot.R
\name{blank.plot}
\alias{blank.plot}
\title{Draw blank plot}
\usage{
blank.plot()
}
\description{
\code{blank.plot} draws a blank plot (no data, axis, or labels) on 
the current device.
}
\details{
Used by the \code{autoimage} function to fill remaining regions 
with white space when there are more plotting regions than images 
to plot.
}
\examples{
blank.plot()
}
\seealso{
\code{autoimage}
}

