% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copoly.R
\docType{data}
\name{copoly}
\alias{copoly}
\title{Colorado state border}
\format{Contains:
\describe{
 \item{x}{longitude coordinates for Colorado border}
 \item{y}{latitude coordinates for Colorado border}
}}
\source{
The \code{\link[maps]{stateMapEnv}} data set in the 
\code{maps} package.
}
\usage{
data(copoly)
}
\description{
A list with components \code{x} and \code{y} specifying
the borders of the state of Colorado in longitude/latitude coordinates.  
This was derived from the \code{\link[maps]{stateMapEnv}} data set 
in the \code{maps} package.
}

