% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_functions_other.R
\name{.legend.scale.args}
\alias{.legend.scale.args}
\title{Creates hidden information for the legend margin and legend scale
to using autolegend function.}
\usage{
.legend.scale.args(new)
}
\description{
Creates hidden information for the legend margin and legend scale
to using autolegend function.
}
\keyword{internal}
