% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inarccap.R
\docType{data}
\name{inarccap}
\alias{inarccap}
\alias{ilon}
\alias{ilat}
\alias{itasmax}
\title{Interpolated maximum daily surface air temperatures on a regular grid.}
\format{
Contains: \describe{ 
\item{ilon}{A vector of of longitude coordinates.} 
\item{ilat}{A vector of latitude coordinates.} 
\item{itasmax}{A 140\eqn{\times}115\eqn{\times}5 array of tasmax values.} }
}
\usage{
data(inarccap)
}
\description{
These data are the \code{narccap} data interpolated onto a regular 
140\eqn{\times}115 grid using the \code{akima::interp} function.
}
\seealso{
\code{\link[autoimage]{narccap}}, \code{akima::interp}
}
