% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppoints.R
\name{ppoints}
\alias{ppoints}
\title{Projected points function}
\usage{
ppoints(x, y = NULL, type = "p", proj, ...)
}
\arguments{
\item{x}{coordinate vectors of points to plot.}

\item{y}{Locations of grid points at which the values in \code{z} 
are measured.  The values must be finite and non-missing.  These 
arguments can be either vectors, in which case the values must be in (strictly) ascending order.  If an irregular grid is to be used, then \code{x} and \code{y} should be numeric matrices having the same number of rows and columns as \code{z}.  If these arguments are not provided, equally spaced values from 0 to 1 are used by default. If \code{x} is a list, its components \code{x$x} and \code{x$y} are used for \code{x} and \code{y}, respectively. If the list has component \code{x$z}, this is used for \code{z}.}

\item{type}{character indicating the type of plotting; actually any of
    the \code{type}s as in \code{\link{plot.default}}.}

\item{proj}{A character string indicating what projection should be
used for the included \code{x} and \code{y} coordinates.  The
default is \code{"none"}.  The other valid choices correspond to the
\code{"projection"} argument in the \code{\link[mapproj]{mapproject}} 
function, which is used for the projection.}

\item{...}{Further graphical parameters passed to the 
\code{\link[graphics]{points}} function.}
}
\description{
\code{ppoints} draws a sequence of points for projected coordinates.
}
\details{
The \code{\link[mapproj]{mapproject}} function is used for projection.
}
\examples{
data(narccap)
# plot image using bonne projection (w/o grid lines)
pimage(lon, lat, tasmax[,,1], proj = "bonne",
       proj.args = list(parameters = 40), 
       paxes.args = list(grid = FALSE))
# get U.S. cities with population of about 40k or more
data(us.cities, package = "maps")
# add cities to existing plot
ppoints(us.cities$long, us.cities$lat, proj = "bonne")
}
\seealso{
\code{\link[graphics]{points}}, \code{\link[mapproj]{mapproject}}, \code{\link[autoimage]{pimage}}
}

