% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psegments.R
\name{psegments}
\alias{psegments}
\title{Projected segments function}
\usage{
psegments(x0, y0, x1 = x0, y1 = y0, proj, ...)
}
\arguments{
\item{x0}{coordinates of points \bold{from} which to draw.}

\item{x1}{coordinates of points \bold{to} which to draw.  At least
    one must be supplied.}

\item{proj}{A character string indicating what projection should be
used for the included \code{x} and \code{y} coordinates.  The
default is \code{"none"}.  The other valid choices correspond to the
\code{"projection"} argument in the \code{\link[mapproj]{mapproject}} 
function, which is used for the projection.}

\item{...}{Additional arguments to pass to the 
\code{\link[graphics]{segments}} function.}
}
\description{
\code{psegments} takes pairs of coordinates and draws line segments
between them, possibly after projection.
}
\details{
The \code{\link[mapproj]{mapproject}} function is used for projection.
}
\examples{
data(narccap)
# plot image using bonne projection (w/o grid lines)
pimage(lon, lat, tasmax[,,1], proj = "bonne",
       proj.args = list(parameters = 40), 
       paxes.args = list(grid = FALSE))
# load some data for larger U.S. cities
data(us.cities, package = "maps")
cityxy <- list(x = us.cities$long[1:5], y = us.cities$lat[1:5])
# plot segments between cities with thick line
psegments(cityxy$x[1:4], cityxy$y[1:4],
          cityxy$x[2:5], cityxy$y[2:5], 
          proj = "bonne", lwd = 3)
}
\seealso{
\code{\link[graphics]{segments}}, \code{\link[mapproj]{mapproject}}, \code{\link[autoimage]{pimage}}
}

