% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.par.R
\name{reset.par}
\alias{reset.par}
\title{Reset par}
\usage{
reset.par()
}
\description{
\code{reset.par} resets the arguments of
\code{\link[graphics]{par}} to the default values when first
opening R (as of version 3.2.2).
}
\examples{
par("mar") #current values of mar
par(mar = c(0, 0, 0, 0)) # change values of mar
par("mar") # changed values of mar
reset.par() # reset to defaults (not necessarilly current values)
par("mar") # should be c(5.1, 4.1, 4.1, 2.1)
}
\seealso{
\code{\link[graphics]{par}}
}

